+++ testCode/testCode/testCode/Base/App/Game/RPG/MVC/Models/PlayerHero.swift
@@ -8,15 +8,24 @@
 
 import Foundation
 
+struct Hero {
+    var name: String
+    var hitPoint: Int
+    let magicPoint: Int
+    let weapons: [Weapon]
+}
+
 class PlayerHero: Monster {
     var hero: Hero = Hero(name: "勇者ヨシヒコ", hitPoint: 90, magicPoint: 12, weapons: [])
     
+    /// プレイヤー：勇者陣 の処理の書き換え
     override init(name: String,hitPoint: Int) {
         hero.name = name
         hero.hitPoint = hitPoint
         super.init(name: name, hitPoint: hitPoint)
     }
     
+    /// プレイヤー：勇者陣 が アイテムをゲットする時の処理
     func getItems ( items: [Item] ) {
         print(items)
     }
